Split up clothing files.
