+++ src/data/clothing/clothing.ts
@@ -2,62 +2,6 @@ import { IItem } from '@/types/item';
 
 const data: IItem[] = [
   {
-    date: '2023-03-07',
-    image:
-      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/unisex-sherpa-denim-jacket-black-denim-front-63694c0e01cf5_1024x1024@2x.jpg',
-    name: 'Denim Embroidered Howler Sherpa Jacket (Black)',
-    tags: ['jacket', 'official'],
-    shop: {
-      name: 'Lit Escalates',
-      url: 'https://www.lit-escalates.com/'
-    },
-    description: 'A denim jacket from Lit Escalates. Pegasus Legion sigil on the back.',
-    link: 'https://www.lit-escalates.com/collections/apparel/products/embroidered-howler-denim-sherpa-jacket'
-  },
-  {
-    date: '2023-03-10',
-    image:
-      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/premium-recycled-bomber-jacket-black-front-636949f435bc0_1024x1024.jpg',
-    name: 'Bomber Howler Embroidered Jacket (Space Black)',
-    tags: ['jacket', 'official'],
-    shop: {
-      name: 'Lit Escalates',
-      url: 'https://www.lit-escalates.com/'
-    },
-    description:
-      'Finally! A place to put your pins and patches. You can never go wrong with a classic bomber jacket, so we made one we thought Sevro would sport.',
-    link: 'https://www.lit-escalates.com/collections/apparel/products/black-howler-bomber-jacket'
-  },
-  {
-    date: '2023-03-10',
-    image:
-      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/unisex-sherpa-denim-jacket-light-denim-left-front-63694e8275d3c_1024x1024.jpg',
-    name: 'Denim Embroidered Howler Sherpa Jacket (Blue)',
-    tags: ['jacket', 'official'],
-    shop: {
-      name: 'Lit Escalates',
-      url: 'https://www.lit-escalates.com/'
-    },
-    description:
-      'Finally! A place to put your pins and patches. Be it the 1990s or mid 3000s—a classic denim jacket never goes out of style.',
-    link: 'https://www.lit-escalates.com/collections/apparel/products/howler-denim-sherpa-jacket'
-  },
-  {
-    date: '2023-03-10',
-    image:
-      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/premium-recycled-bomber-jacket-army-front-2-63693e9b5f242_1024x1024.jpg',
-    name: 'Bomber Howler Embroidered Jacket (Army Green)',
-    tags: ['jacket', 'official'],
-    shop: {
-      name: 'Lit Escalates',
-      url: 'https://www.lit-escalates.com/'
-    },
-    description:
-      'Finally! A place to put your pins and patches. You can never go wrong with a classic bomber jacket, so we made one we thought Sevro would sport.',
-    link: 'https://www.lit-escalates.com/collections/apparel/products/green-howler-bomber-jacket'
-  },
-
-  {
     date: '2023-03-13',
     image:
       'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/unisex-basic-softstyle-t-shirt-black-5ff0df5130a61_1024x1024.jpg',
@@ -101,20 +45,6 @@ const data: IItem[] = [
   {
     date: '2023-03-13',
     image:
-      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/unisex-premium-tank-top-white-front-60e7265b5ea21_1024x1024.jpg',
-    name: 'Omnis Vir Lupus Tank Top',
-    tags: ['shirt', 'official'],
-    shop: {
-      name: 'Lit Escalates',
-      url: 'https://www.lit-escalates.com/'
-    },
-    description:
-      "Ya'll loved the other PB Doodles pieces so much we had to release this one as a tank for summer. A classic, all-purpose unisex tank of course.",
-    link: 'https://www.lit-escalates.com/collections/apparel/products/omnis-vir-lupus-unisex-tank-top'
-  },
-  {
-    date: '2023-03-13',
-    image:
       'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/unisex-basic-softstyle-t-shirt-black-front-61fed4cb71be7_1024x1024.jpg',
     name: 'Helldivers Union T-Shirt',
     tags: ['shirt', 'official'],
@@ -154,57 +84,6 @@ const data: IItem[] = [
     link: 'https://www.lit-escalates.com/collections/apparel/products/red-sigil-t-shirt'
   },
   {
-    date: '2023-03-13',
-    image:
-      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/toddler-jersey-t-shirt-light-blue-front-60e4df6b9d271_1024x1024.jpg',
-    name: 'Mangy Angel Toddler T-shirt',
-    tags: ['shirt', 'morning-star', 'official'],
-    shop: {
-      name: 'Lit Escalates',
-      url: 'https://www.lit-escalates.com/'
-    },
-    description:
-      "For your own Mangy Little Guardian Angel. Whether worn alone or layered, this jersey t-shirt is soft yet durable to withstand your toddler's playtime demands.",
-    link: 'https://www.lit-escalates.com/collections/apparel/products/mangy-angel-toddler-jersey-t-shirt'
-  },
-  {
-    date: '2023-03-13',
-    image:
-      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/youth-jersey-t-shirt-light-blue-front-2-60e4e6f716460_1024x1024.jpg',
-    name: 'Mangy Angel Youth T-shirt',
-    tags: ['shirt', 'morning-star', 'official'],
-    shop: {
-      name: 'Lit Escalates',
-      url: 'https://www.lit-escalates.com/'
-    },
-    description: 'For your very own Mangy Little Guardian Angel.',
-    link: 'https://www.lit-escalates.com/collections/apparel/products/mangy-angel-youth-t-shirt'
-  },
-  {
-    date: '2023-03-14',
-    image: 'https://i.etsystatic.com/16046340/r/il/98620d/4672320544/il_794xN.4672320544_bil8.jpg',
-    name: 'Littlest Howler Toddler Tee',
-    tags: ['shirt'],
-    shop: {
-      name: 'CreativeLeighCrafts',
-      url: 'https://www.etsy.com/shop/CreativeLeighCrafts'
-    },
-    description: 'Perfect tee for your little howler!',
-    link: 'https://www.etsy.com/listing/1413172482/littlest-howler-toddler-short-sleeve-tee'
-  },
-  {
-    date: '2023-03-14',
-    image: 'https://i.etsystatic.com/16046340/r/il/43a4cc/4720357577/il_794xN.4720357577_m4ui.jpg',
-    name: 'Littlest Howler Toddler T-shirt',
-    tags: ['shirt'],
-    shop: {
-      name: 'CreativeLeighCrafts',
-      url: 'https://www.etsy.com/shop/CreativeLeighCrafts'
-    },
-    description: 'Perfect tee for your little howler!',
-    link: 'https://www.etsy.com/listing/1427335795/littlest-howler-toddler-t-shirt'
-  },
-  {
     date: '2023-03-14',
     image: 'https://i.etsystatic.com/16046340/r/il/fd3e1a/4628213168/il_794xN.4628213168_5noy.jpg',
     name: 'Howler Life - Unisex Jersey Tee',
@@ -239,30 +118,6 @@ const data: IItem[] = [
     },
     description: 'Represent your favorite podcast with this HowlerPod Logo T-Shirt.',
     link: 'https://www.etsy.com/listing/867027834/howlerpod-logo-short-sleeve-unisex-t'
-  },
-  {
-    date: '2023-03-15',
-    image: 'https://i.etsystatic.com/21279369/r/il/29c4ea/4400194212/il_1140xN.4400194212_37xl.jpg',
-    name: "Hic Est Lupus Motherfucker Women's Tank",
-    tags: ['shirt'],
-    shop: {
-      name: 'TheMightyPenBookTees',
-      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
-    },
-    description: 'Hic Est Lupus Motherfucker',
-    link: 'https://www.etsy.com/listing/1360171141/hic-est-lupus-explicit-womens-tank'
-  },
-  {
-    date: '2023-03-15',
-    image: 'https://i.etsystatic.com/21279369/r/il/071af0/4447608247/il_1588xN.4447608247_529x.jpg',
-    name: 'Hic Est Lupus Motherfucker Tank Top',
-    tags: ['shirt'],
-    shop: {
-      name: 'TheMightyPenBookTees',
-      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
-    },
-    description: 'Hic Est Lupus Motherfucker.',
-    link: 'https://www.etsy.com/listing/1346190234/hic-est-lupus-explicit-red-rising-unisex'
   },
   {
     date: '2023-03-15',
@@ -339,18 +194,6 @@ const data: IItem[] = [
     description:
       "Hic Sunt Leones! It's a full print House Augustus fan made Red Rising tee - LIMITED GOLD RUSH EDITION.",
     link: 'https://www.etsy.com/listing/1046095364/limited-gold-rush-hic-sunt-leones-au'
-  },
-  {
-    date: '2023-03-15',
-    image: 'https://i.etsystatic.com/21279369/r/il/c3657e/2923200748/il_794xN.2923200748_2e9l.jpg',
-    name: 'HWLR Logo Gold Tank Top',
-    tags: ['shirt'],
-    shop: {
-      name: 'TheMightyPenBookTees',
-      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
-    },
-    description: 'A classic, all-purpose unisex tank with the HWLR logo.',
-    link: 'https://www.etsy.com/listing/976054951/hwlr-logo-gold-unisex-tank-top'
   },
   {
     date: '2023-03-15',
@@ -646,20 +489,6 @@ const data: IItem[] = [
   },
   {
     date: '2023-03-15',
-    image: 'https://i.etsystatic.com/21279369/r/il/3a1652/2869582101/il_1140xN.2869582101_b70t.jpg',
-    name: 'Sleepy Sophocles Baby One Piece',
-    tags: ['shirt', 'sophocles'],
-    shop: {
-      name: 'TheMightyPenBookTees',
-
-      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
-    },
-    description:
-      "Dress your 'lil jelly bean in this Red Rising inspired baby Sophocles 100% cotton one piece.",
-    link: 'https://www.etsy.com/listing/949696559/sleepy-sophocles-red-rising-baby-short'
-  },
-  {
-    date: '2023-03-15',
     image: 'https://i.etsystatic.com/21279369/r/il/feecdd/2869599103/il_1140xN.2869599103_hlln.jpg',
     name: 'Widdle Ares and the Sunblood Unicorn Baby One Piece',
     tags: ['shirt', 'sevro'],
@@ -807,54 +636,6 @@ const data: IItem[] = [
     link: 'https://www.etsy.com/listing/1384125202/red-rising-victra-bookish-valentines-day'
   },
   {
-    date: '2023-03-21',
-    image: 'https://i.etsystatic.com/10165646/r/il/8c158d/4566275456/il_1140xN.4566275456_feip.jpg',
-    name: 'Omnis Vir Lupus Toddler Tee',
-    tags: ['shirt'],
-    shop: {
-      name: 'LiveForMore',
-      url: 'https://www.etsy.com/shop/LiveForMore'
-    },
-    description: 'Omnis vir Lupus design on a toddler tee.',
-    link: 'https://www.etsy.com/listing/1403976331/omnis-vir-lupus-red-rising-howler'
-  },
-  {
-    date: '2023-03-21',
-    image: 'https://i.etsystatic.com/10165646/r/il/15825a/4614598903/il_794xN.4614598903_h8sk.jpg',
-    name: 'Omnis Vir Lupus Youth Tee',
-    tags: ['shirt'],
-    shop: {
-      name: 'LiveForMore',
-      url: 'https://www.etsy.com/shop/LiveForMore'
-    },
-    description: "Omnis vir Lupus design on a kid's tee.",
-    link: 'https://www.etsy.com/listing/1389758824/omnis-vir-lupus-red-rising-kid-shirt'
-  },
-  {
-    date: '2023-03-21',
-    image: 'https://i.etsystatic.com/10165646/r/il/86b3d3/4566205010/il_794xN.4566205010_grx4.jpg',
-    name: 'Omnis Vir Lupus Youth Tee',
-    tags: ['shirt'],
-    shop: {
-      name: 'LiveForMore',
-      url: 'https://www.etsy.com/shop/LiveForMore'
-    },
-    description: "Omnis vir Lupus design on a kid's tee.",
-    link: 'https://www.etsy.com/listing/1389746156/omnis-vir-lupus-red-rising-kid-shirt'
-  },
-  {
-    date: '2023-03-21',
-    image: 'https://i.etsystatic.com/10165646/r/il/711281/4614432879/il_1140xN.4614432879_bd3q.jpg',
-    name: 'Omnis Vir Lupus Youth Tee',
-    tags: ['shirt'],
-    shop: {
-      name: 'LiveForMore',
-      url: 'https://www.etsy.com/shop/LiveForMore'
-    },
-    description: "Omnis vir Lupus design on a kid's tee.",
-    link: 'https://www.etsy.com/listing/1403955553/omnis-vir-lupus-red-rising-howler'
-  },
-  {
     date: '2023-03-22',
     image: 'https://i.etsystatic.com/23884463/r/il/0402f9/2973735309/il_794xN.2973735309_m434.jpg',
     name: 'Son of Ares Shirt',
@@ -918,19 +699,6 @@ const data: IItem[] = [
     link: 'https://www.etsy.com/listing/1383421762/a-sage-digs-the-roots-red-rising-lorn-au'
   },
   {
-    date: '2023-03-29',
-    image: 'https://i.etsystatic.com/36119472/r/il/3f3902/4539504926/il_794xN.4539504926_ts4h.jpg',
-    name: 'A Sage Digs The Roots Tank',
-    tags: ['shirt'],
-    shop: {
-      name: 'TooManyBones',
-      url: 'https://www.etsy.com/shop/TooManyBones'
-    },
-    description:
-      'Suns out guns out! Take your love for the world of Pierce Brown out in the world in maximum comfort!',
-    link: 'https://www.etsy.com/listing/1383453636/fantasy-tank-top-a-sage-digs-the-roots'
-  },
-  {
     date: '2023-03-30',
     image: 'https://img.teeherivar.com/200514AFD6052-1.jpg',
     name: 'The Howlers Shirt',
@@ -957,8 +725,8 @@ const data: IItem[] = [
   {
     date: '2023-03-30',
     image: 'https://i.etsystatic.com/18738093/r/il/bfd5ce/4606134316/il_794xN.4606134316_jt6a.jpg',
-    name: 'Red Rising Shirt',
-    tags: ['shirt'],
+    name: 'Red Rising (Sweat)Shirt',
+    tags: ['shirt', 'sweatshirt'],
     shop: {
       name: 'BertieandMo',
       url: 'https://www.etsy.com/shop/BertieandMo'
