+++ src/data/clothing/hats.ts
