+++ src/data/clothing/jackets.ts
@@ -1,0 +1,61 @@
+import { IItem } from '@/types/item';
+
+const data: IItem[] = [
+  {
+    date: '2023-03-07',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/unisex-sherpa-denim-jacket-black-denim-front-63694c0e01cf5_1024x1024@2x.jpg',
+    name: 'Denim Embroidered Howler Sherpa Jacket (Black)',
+    tags: ['jacket', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description: 'A denim jacket from Lit Escalates. Pegasus Legion sigil on the back.',
+    link: 'https://www.lit-escalates.com/collections/apparel/products/embroidered-howler-denim-sherpa-jacket'
+  },
+  {
+    date: '2023-03-10',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/premium-recycled-bomber-jacket-black-front-636949f435bc0_1024x1024.jpg',
+    name: 'Bomber Howler Embroidered Jacket (Space Black)',
+    tags: ['jacket', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description:
+      'Finally! A place to put your pins and patches. You can never go wrong with a classic bomber jacket, so we made one we thought Sevro would sport.',
+    link: 'https://www.lit-escalates.com/collections/apparel/products/black-howler-bomber-jacket'
+  },
+  {
+    date: '2023-03-10',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/unisex-sherpa-denim-jacket-light-denim-left-front-63694e8275d3c_1024x1024.jpg',
+    name: 'Denim Embroidered Howler Sherpa Jacket (Blue)',
+    tags: ['jacket', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description:
+      'Finally! A place to put your pins and patches. Be it the 1990s or mid 3000s—a classic denim jacket never goes out of style.',
+    link: 'https://www.lit-escalates.com/collections/apparel/products/howler-denim-sherpa-jacket'
+  },
+  {
+    date: '2023-03-10',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/premium-recycled-bomber-jacket-army-front-2-63693e9b5f242_1024x1024.jpg',
+    name: 'Bomber Howler Embroidered Jacket (Army Green)',
+    tags: ['jacket', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description:
+      'Finally! A place to put your pins and patches. You can never go wrong with a classic bomber jacket, so we made one we thought Sevro would sport.',
+    link: 'https://www.lit-escalates.com/collections/apparel/products/green-howler-bomber-jacket'
+  }
+];
+
+export default data;
