+++ src/data/clothing/kids.ts
@@ -1,0 +1,143 @@
+import { IItem } from '@/types/item';
+
+const data: IItem[] = [
+  {
+    date: '2023-03-21',
+    image: 'https://i.etsystatic.com/10165646/r/il/15825a/4614598903/il_794xN.4614598903_h8sk.jpg',
+    name: 'Omnis Vir Lupus Youth Tee',
+    tags: ['shirt'],
+    shop: {
+      name: 'LiveForMore',
+      url: 'https://www.etsy.com/shop/LiveForMore'
+    },
+    description: "Omnis vir Lupus design on a kid's tee.",
+    link: 'https://www.etsy.com/listing/1389758824/omnis-vir-lupus-red-rising-kid-shirt'
+  },
+  {
+    date: '2023-03-21',
+    image: 'https://i.etsystatic.com/10165646/r/il/86b3d3/4566205010/il_794xN.4566205010_grx4.jpg',
+    name: 'Omnis Vir Lupus Youth Tee',
+    tags: ['shirt'],
+    shop: {
+      name: 'LiveForMore',
+      url: 'https://www.etsy.com/shop/LiveForMore'
+    },
+    description: "Omnis vir Lupus design on a kid's tee.",
+    link: 'https://www.etsy.com/listing/1389746156/omnis-vir-lupus-red-rising-kid-shirt'
+  },
+  {
+    date: '2023-03-21',
+    image: 'https://i.etsystatic.com/10165646/r/il/711281/4614432879/il_1140xN.4614432879_bd3q.jpg',
+    name: 'Omnis Vir Lupus Youth Tee',
+    tags: ['shirt'],
+    shop: {
+      name: 'LiveForMore',
+      url: 'https://www.etsy.com/shop/LiveForMore'
+    },
+    description: "Omnis vir Lupus design on a kid's tee.",
+    link: 'https://www.etsy.com/listing/1403955553/omnis-vir-lupus-red-rising-howler'
+  },
+  {
+    date: '2023-03-13',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/youth-jersey-t-shirt-light-blue-front-2-60e4e6f716460_1024x1024.jpg',
+    name: 'Mangy Angel Youth T-shirt',
+    tags: ['shirt', 'morning-star', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description: 'For your very own Mangy Little Guardian Angel.',
+    link: 'https://www.lit-escalates.com/collections/apparel/products/mangy-angel-youth-t-shirt'
+  },
+  {
+    date: '2023-03-13',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/toddler-jersey-t-shirt-light-blue-front-60e4df6b9d271_1024x1024.jpg',
+    name: 'Mangy Angel Toddler T-shirt',
+    tags: ['shirt', 'morning-star', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description:
+      "For your own Mangy Little Guardian Angel. Whether worn alone or layered, this jersey t-shirt is soft yet durable to withstand your toddler's playtime demands.",
+    link: 'https://www.lit-escalates.com/collections/apparel/products/mangy-angel-toddler-jersey-t-shirt'
+  },
+  {
+    date: '2023-03-14',
+    image: 'https://i.etsystatic.com/16046340/r/il/98620d/4672320544/il_794xN.4672320544_bil8.jpg',
+    name: 'Littlest Howler Toddler Tee',
+    tags: ['shirt'],
+    shop: {
+      name: 'CreativeLeighCrafts',
+      url: 'https://www.etsy.com/shop/CreativeLeighCrafts'
+    },
+    description: 'Perfect tee for your little howler!',
+    link: 'https://www.etsy.com/listing/1413172482/littlest-howler-toddler-short-sleeve-tee'
+  },
+  {
+    date: '2023-03-14',
+    image: 'https://i.etsystatic.com/16046340/r/il/43a4cc/4720357577/il_794xN.4720357577_m4ui.jpg',
+    name: 'Littlest Howler Toddler T-shirt',
+    tags: ['shirt'],
+    shop: {
+      name: 'CreativeLeighCrafts',
+      url: 'https://www.etsy.com/shop/CreativeLeighCrafts'
+    },
+    description: 'Perfect tee for your little howler!',
+    link: 'https://www.etsy.com/listing/1427335795/littlest-howler-toddler-t-shirt'
+  },
+  {
+    date: '2023-03-21',
+    image: 'https://i.etsystatic.com/10165646/r/il/8c158d/4566275456/il_1140xN.4566275456_feip.jpg',
+    name: 'Omnis Vir Lupus Toddler Tee',
+    tags: ['shirt'],
+    shop: {
+      name: 'LiveForMore',
+      url: 'https://www.etsy.com/shop/LiveForMore'
+    },
+    description: 'Omnis vir Lupus design on a toddler tee.',
+    link: 'https://www.etsy.com/listing/1403976331/omnis-vir-lupus-red-rising-howler'
+  },
+  {
+    date: '2023-03-21',
+    image: 'https://i.etsystatic.com/10165646/r/il/8dd3fb/4614481191/il_794xN.4614481191_3gpt.jpg',
+    name: 'Omnis Vir Lupus Toddler Hoodie',
+    tags: ['hoodie'],
+    shop: {
+      name: 'LiveForMore',
+      url: 'https://www.etsy.com/shop/LiveForMore'
+    },
+    description: 'Omnis vir Lupus design on a toddler hoodie.',
+    link: 'https://www.etsy.com/listing/1403966849/red-rising-omnis-vir-lupus-toddler'
+  },
+  {
+    date: '2023-03-21',
+    image: 'https://i.etsystatic.com/10165646/r/il/80d25f/4614374225/il_1140xN.4614374225_egt7.jpg',
+    name: 'Omnis Vir Lupus Toddler Hoodie',
+    tags: ['hoodie'],
+    shop: {
+      name: 'LiveForMore',
+      url: 'https://www.etsy.com/shop/LiveForMore'
+    },
+    description: 'Omnis vir Lupus design on a toddler hoodie.',
+    link: 'https://www.etsy.com/listing/1403942321/red-rising-omnis-vir-lupus-toddler'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/3a1652/2869582101/il_1140xN.2869582101_b70t.jpg',
+    name: 'Sleepy Sophocles Baby One Piece',
+    tags: ['shirt', 'sophocles'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description:
+      "Dress your 'lil jelly bean in this Red Rising inspired baby Sophocles 100% cotton one piece.",
+    link: 'https://www.etsy.com/listing/949696559/sleepy-sophocles-red-rising-baby-short'
+  }
+];
+
+export default data;
