+++ src/data/clothing/tanks.ts
@@ -1,0 +1,69 @@
+import { IItem } from '@/types/item';
+
+const data: IItem[] = [
+  {
+    date: '2023-03-13',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/unisex-premium-tank-top-white-front-60e7265b5ea21_1024x1024.jpg',
+    name: 'Omnis Vir Lupus Tank Top',
+    tags: ['shirt', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description:
+      "Ya'll loved the other PB Doodles pieces so much we had to release this one as a tank for summer. A classic, all-purpose unisex tank of course.",
+    link: 'https://www.lit-escalates.com/collections/apparel/products/omnis-vir-lupus-unisex-tank-top'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/29c4ea/4400194212/il_1140xN.4400194212_37xl.jpg',
+    name: "Hic Est Lupus Motherfucker Women's Tank",
+    tags: ['shirt'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'Hic Est Lupus Motherfucker',
+    link: 'https://www.etsy.com/listing/1360171141/hic-est-lupus-explicit-womens-tank'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/071af0/4447608247/il_1588xN.4447608247_529x.jpg',
+    name: 'Hic Est Lupus Motherfucker Tank Top',
+    tags: ['shirt'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'Hic Est Lupus Motherfucker.',
+    link: 'https://www.etsy.com/listing/1346190234/hic-est-lupus-explicit-red-rising-unisex'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/c3657e/2923200748/il_794xN.2923200748_2e9l.jpg',
+    name: 'HWLR Logo Gold Tank Top',
+    tags: ['shirt'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'A classic, all-purpose unisex tank with the HWLR logo.',
+    link: 'https://www.etsy.com/listing/976054951/hwlr-logo-gold-unisex-tank-top'
+  },
+  {
+    date: '2023-03-29',
+    image: 'https://i.etsystatic.com/36119472/r/il/3f3902/4539504926/il_794xN.4539504926_ts4h.jpg',
+    name: 'A Sage Digs The Roots Tank',
+    tags: ['shirt'],
+    shop: {
+      name: 'TooManyBones',
+      url: 'https://www.etsy.com/shop/TooManyBones'
+    },
+    description:
+      'Suns out guns out! Take your love for the world of Pierce Brown out in the world in maximum comfort!',
+    link: 'https://www.etsy.com/listing/1383453636/fantasy-tank-top-a-sage-digs-the-roots'
+  }
+];
+
+export default data;
