+++ src/data/data.ts
@@ -1,14 +1,10 @@
 import bookmarks from './bookmarks';
 import books from './books';
 import candles from './candles';
-import clothing from './clothing';
 import digital from './digital';
 import discs from './discs';
-import hats from './hats';
-import hoodies from './hoodies';
 import jewelry from './jewelry';
 import magnets from './magnets';
-import mugs from './mugs';
 import notebooks from './notebooks';
 import other from './other';
 import patches from './patches';
@@ -16,30 +12,40 @@ import phoneCases from './phone-cases';
 import pins from './pins';
 import prints from './prints';
 import stickers from './stickers';
-import sweatshirts from './sweatshirts';
+
+import clothing from './clothing/clothing';
+import hats from './clothing/hats';
+import hoodies from './clothing/hoodies';
+import jackets from './clothing/jackets';
+import kids from './clothing/kids';
+import masks from './clothing/masks';
+import sweatshirts from './clothing/sweatshirts';
+import tanks from './clothing/tanks';
 
 import drinkware from './drinkware/drinkware';
+import mugs from './drinkware/mugs';
 import waterBottles from './drinkware/water-bottles';
 
 export const ALL_DATA = [
+  [...drinkware, ...waterBottles],
+  [...clothing, ...jackets, ...kids, ...tanks],
   ...bookmarks,
   ...books,
   ...candles,
-  ...clothing,
   ...digital,
   ...discs,
   ...hats,
   ...hoodies,
   ...jewelry,
   ...magnets,
-  ...notebooks,
+  ...masks,
   ...mugs,
+  ...notebooks,
   ...other,
   ...patches,
   ...phoneCases,
   ...pins,
   ...prints,
   ...stickers,
-  ...sweatshirts,
-  [...drinkware, ...waterBottles]
+  ...sweatshirts
 ].flat();
